apps/ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "psycopg2-binary==2.9.10",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "yfinance>=0.2.61",
    "pandas>=2.2.3",
    "numpy>=1.26.0",
]
